Fix more unit tests
